Conflicts:
	src/main/java/net/bither/qrcode/SelectTransportQRCodePanel.java

+++ src/main/java/net/bither/db/PeerProvider.java
@@ -24,7 +24,10 @@ import net.bither.bitherj.utils.Utils;
 
 import java.net.InetAddress;
 import java.net.UnknownHostException;
-import java.sql.*;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -34,7 +37,7 @@ public class PeerProvider implements IPeerProvider {
             "(peer_address,peer_port,peer_services,peer_timestamp,peer_connected_cnt)" +
             " values (?,?,?,?,?) ";
 
-    private static PeerProvider peerProvider = new PeerProvider(ApplicationInstanceManager.mDBHelper);
+    private static PeerProvider peerProvider = new PeerProvider(ApplicationInstanceManager.txDBHelper);
 
     public static PeerProvider getInstance() {
         return peerProvider;
@@ -50,7 +53,7 @@ public class PeerProvider implements IPeerProvider {
         List<Peer> peers = new ArrayList<Peer>();
         String sql = "select * from peers";
         try {
-            ResultSet c = this.mDb.query(sql);
+            ResultSet c = this.mDb.query(sql, null);
             while (c.next()) {
                 Peer peer = applyCursor(c);
                 if (peer != null) {
@@ -68,7 +71,7 @@ public class PeerProvider implements IPeerProvider {
     public void deletePeersNotInAddresses(List<InetAddress> peerAddrsses) {
         final List<Long> needDeletePeers = new ArrayList<Long>();
         String sql = "select peer_address from peers";
-        ResultSet c = this.mDb.query(sql);
+        ResultSet c = this.mDb.query(sql, null);
         try {
             while (c.next()) {
                 int idColumn = c.findColumn(AbstractDb.PeersColumns.PEER_ADDRESS);
@@ -91,20 +94,18 @@ public class PeerProvider implements IPeerProvider {
         } catch (SQLException e) {
             e.printStackTrace();
         }
-        this.mDb.executeUpdate(new BitherDBHelper.IExecuteDB() {
-            @Override
-            public void execute(Connection conn) {
-                try {
-                    for (long i : needDeletePeers) {
-                        PreparedStatement preparedStatement = conn.prepareStatement("delete peers where peer_address=?");
-                        preparedStatement.setLong(1, i);
 
-                    }
-                } catch (SQLException e) {
-                    e.printStackTrace();
-                }
+        try {
+            this.mDb.getConn().setAutoCommit(false);
+            for (long i : needDeletePeers) {
+                PreparedStatement preparedStatement = this.mDb.getConn().prepareStatement("delete peers where peer_address=?");
+                preparedStatement.setLong(1, i);
+
             }
-        });
+            this.mDb.getConn().commit();
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
 
 
     }
@@ -130,38 +131,46 @@ public class PeerProvider implements IPeerProvider {
             }
         }
         if (addItems.size() > 0) {
-            this.mDb.executeUpdate(new BitherDBHelper.IExecuteDB() {
-                @Override
-                public void execute(Connection conn) throws SQLException {
-                    for (Peer item : addItems) {
-                        PreparedStatement preparedStatement = conn.prepareStatement(insertPeerSql);
-                        preparedStatement.setLong(1, Utils.parseLongFromAddress(item
-                                .getPeerAddress()));
-                        preparedStatement.setLong(2, item.getPeerPort());
-                        preparedStatement.setLong(3, item.getPeerServices());
-                        preparedStatement.setLong(4, item.getPeerTimestamp());
-                        preparedStatement.setLong(5, item.getPeerConnectedCnt());
-                        preparedStatement.executeUpdate();
-                    }
+            try {
+                this.mDb.getConn().setAutoCommit(false);
+                for (Peer item : addItems) {
+                    PreparedStatement preparedStatement = this.mDb.getConn().prepareStatement(insertPeerSql);
+                    preparedStatement.setLong(1, Utils.parseLongFromAddress(item
+                            .getPeerAddress()));
+                    preparedStatement.setLong(2, item.getPeerPort());
+                    preparedStatement.setLong(3, item.getPeerServices());
+                    preparedStatement.setLong(4, item.getPeerTimestamp());
+                    preparedStatement.setLong(5, item.getPeerConnectedCnt());
+                    preparedStatement.executeUpdate();
                 }
-            });
+                this.mDb.getConn().commit();
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
 
         }
     }
 
 
     public void removePeer(InetAddress address) {
-        this.mDb.executeUpdate("delete from peers where peer_address = ?", new String[]{Long.toString(Utils.parseLongFromAddress
-                (address))});
+        try {
+            if (this.mDb.getConn() != null && !this.mDb.getConn().isClosed()) {
+                this.mDb.executeUpdate("delete from peers where peer_address = ?", new String[]{Long.toString(Utils.parseLongFromAddress
+                        (address))});
+
+//TODO Database synchronization is wrong
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
     }
 
     public void conncetFail(InetAddress address) {
         try {
             long addressLong = Utils.parseLongFromAddress(address);
-            String sql = "select count(0) cnt from peers where peer_address=" + Long.toString
-                    (addressLong) + " and peer_connected_cnt=0";
+            String sql = "select count(0) cnt from peers where peer_address=? and peer_connected_cnt=0";
 
-            ResultSet c = this.mDb.query(sql);
+            ResultSet c = this.mDb.query(sql, new String[]{Long.toString(addressLong)});
             int cnt = 0;
             if (c.next()) {
                 int idColumn = c.findColumn("cnt");
@@ -194,9 +203,9 @@ public class PeerProvider implements IPeerProvider {
 
     public List<Peer> getPeersWithLimit(int limit) {
         List<Peer> peerItemList = new ArrayList<Peer>();
-        String sql = "select * from peers order by peer_address limit " + Integer.toString(limit);
+        String sql = "select * from peers order by peer_address limit ?";
         try {
-            ResultSet c = this.mDb.query(sql);
+            ResultSet c = this.mDb.query(sql, new String[]{Integer.toString(limit)});
             while (c.next()) {
                 Peer peer = applyCursor(c);
                 if (peer != null) {
@@ -219,7 +228,7 @@ public class PeerProvider implements IPeerProvider {
                     "peer_connected_cnt<>1";
             int disconnectingPeerCnt = 0;
 
-            ResultSet c = this.mDb.query(disconnectingPeerCntSql);
+            ResultSet c = this.mDb.query(disconnectingPeerCntSql, null);
             if (c.next()) {
                 int idColumn = c.findColumn("cnt");
                 if (idColumn != -1) {
@@ -229,9 +238,8 @@ public class PeerProvider implements IPeerProvider {
             c.close();
             if (disconnectingPeerCnt > maxPeerSaveCnt) {
                 String sql = "select peer_timestamp from peers where peer_connected_cnt<>1 " +
-                        "order by peer_timestamp desc limit 1 offset " + Integer.toString
-                        (maxPeerSaveCnt);
-                c = this.mDb.query(sql);
+                        "order by peer_timestamp desc limit 1 offset ? ";
+                c = this.mDb.query(sql, new String[]{Integer.toString(maxPeerSaveCnt)});
                 long timestamp = 0;
                 if (c.next()) {
                     int idColumn = c.findColumn(AbstractDb.PeersColumns.PEER_TIMESTAMP);
@@ -293,16 +301,15 @@ public class PeerProvider implements IPeerProvider {
     }
 
     public void recreate() {
-        this.mDb.executeUpdate(new BitherDBHelper.IExecuteDB() {
-            @Override
-            public void execute(Connection conn) throws SQLException {
-                Statement stmt = conn.createStatement();
-                stmt.executeUpdate("drop table " + AbstractDb.Tables.PEERS + ";");
-                stmt.executeUpdate(AbstractDb.CREATE_PEER_SQL);
-                conn.commit();
-            }
-        });
+        try {
+            Statement stmt = this.mDb.getConn().createStatement();
+            stmt.executeUpdate("drop table " + AbstractDb.Tables.PEERS + ";");
+            stmt.executeUpdate(AbstractDb.CREATE_PEER_SQL);
+            this.mDb.getConn().commit();
 
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
     }
 
 }
